fix(shell): fix all errors in validation shells

Former-commit-id: a97e919a9538a9dabab44310c888e18a8579d71d

+++ shellScript/validation.sh
@@ -2,10 +2,14 @@
 
 workers=("2.2.2.101" "2.2.2.102" "2.2.2.104" "2.2.2.105" "2.2.2.106" "2.2.2.107" "2.2.2.108" "2.2.2.109" "2.2.2.110" "2.2.2.111")
 name="$1"
+[ -z "$name" ] && { echo ">> error: No name provided." >&2; exit 1; }
 home_folder="/home/green"
+
 remote_folder="$home_folder/dataset/${name}_output"
-master_local_folder="$home_folder/validation"
-output_file="$home_folder/validation/merged_file"
+remote_save_folder="$home_folder/validation/${name}"
+
+master_local_folder="$home_folder/validation/headtail"
+output_file="$home_folder/validation/validation_merged_file.txt"
 
 set -e
 rm -r "$master_local_folder" || true
@@ -24,17 +28,20 @@ for worker in "${workers[@]}"; do
         done
     "
     
+    ssh "$worker" "rm -r \"$remove_save_folder\" || true"
+    ssh "$worker" "mkdir -p \"$remote_save_folder\""
+
     # extract head & tail from files
     ssh "$worker" "
         for FILE in $remote_folder/*; do
             BASENAME=\$(basename \"\$FILE\")
-            head -n 1 \"\$FILE\" > \"$remote_folder/headtail_\$BASENAME.txt\"
-            tail -n 1 \"\$FILE\" >> \"$remote_folder/headtail_\$BASENAME.txt\"
+            head -n 1 \"\$FILE\" > \"$remote_save_folder/headtail_\$BASENAME.txt\"
+            tail -n 1 \"\$FILE\" >> \"$remote_save_folder/headtail_\$BASENAME.txt\"
         done
     "
 
     # send head & tail to master
-    scp "$worker:$remote_folder/headtail_*" "$master_local_folder/"
+    scp "$worker:$remote_save_folder/headtail_*" "$master_local_folder/"
     
     echo " "
 done
